+++ lib/ansible/modules/cloud/alicloud/ali_slb_lb_facts.py
@@ -24,133 +24,224 @@ ANSIBLE_METADATA = {'metadata_version': '1.1',
                     'status': ['preview'],
                     'supported_by': 'community'}
 
-DOCUMENTATION = """
+DOCUMENTATION = '''
 ---
 module: ali_slb_lb_facts
-version_added: "1.5.0"
+version_added: "2.8"
 short_description: Gather facts on server load balancer of Alibaba Cloud.
 description:
      - This module fetches data from the Open API in Alicloud.
        The module must be called from within the SLB itself.
 
 options:
-    load_balancer_name:
-      description:
-        - A list of server laod balancer names.
-      aliases: [ "name"]
-    load_balancer_ids:
-      description:
-        - A list of server load balancer ids.
-      aliases: [ "ids" ]
+  load_balancer_name:
+    description:
+      - (Deprecated) A list of server laod balancer names. New option `name_prefix` instead.
+    aliases: ["name"]
+  load_balancer_ids:
+    description:
+      - A list of load balancer IDs to gather facts for.
+    aliases: ['ids']
+  name_prefix:
+    description:
+      - Use a load balancer name prefix to filter load balancers.
+  filters:
+    description:
+      - A dict of filters to apply. Each dict item consists of a filter key and a filter value. The filter keys can be
+        all of request parameters. See U(https://www.alibabacloud.com/help/doc-detail/27582.htm) for parameter details.
+        Filter keys can be same as request parameter name or be lower case and use underscores ("_") or dashes ("-") to
+        connect different words in one parameter. 'LoadBalancerId' will be appended to I(load_balancer_ids) automatically.
 author:
     - "He Guimin (@xiaozhu36)"
 requirements:
     - "python >= 2.6"
-    - "footmark"
+    - "footmark >= 1.9.0"
 extends_documentation_fragment:
     - alicloud
-"""
+'''
 
 EXAMPLES = '''
-# Fetch server load balancers according to setting different filters
-- name: Fetch server load balancer example
-  hosts: localhost
-  connection: local
-  vars:   
-    alicloud_access_key: <your-alicloud-access-key-id>
-    alicloud_secret_key: <your-alicloud-access-secret-key>
-    alicloud_region: cn-beijing
-    load_balancer_ids:
-      - lb-dj1tlosv0oqbkho4bitj4
-      - lb-dj1e5kwh41n87vkn1pxn5
-    load_balancer_name: 
-          - test1
-          - test
-  tasks:
-    - name: Find all server load balancers in specified region.
-      ali_slb_lb_facts:
-        alicloud_access_key: '{{ alicloud_access_key }}'
-        alicloud_secret_key: '{{ alicloud_secret_key }}'
-        alicloud_region: '{{ alicloud_region }}'        
-      register: all_slb
-    - debug: var=all_slb
+# Note: These examples do not set authentication details, see the Alibaba Cloud Guide for details.
+- name: Retrieving slbs using ids
+  ali_slb_lb_facts:
+    ids: 'lb-sn33f3'
 
-    - name: Find all server load balancers by ids
-      ali_slb_lb_facts:
-        alicloud_access_key: '{{ alicloud_access_key }}'
-        alicloud_secret_key: '{{ alicloud_secret_key }}'
-        alicloud_region: '{{ alicloud_region }}'
-        load_balancer_ids: '{{ load_balancer_ids }}'
-      register: slb_by_id
-    - debug: var=slb_by_id
-    
-    - name: Find all server load balancers by name
-      ali_slb_lb_facts:
-        alicloud_access_key: '{{ alicloud_access_key }}'
-        alicloud_secret_key: '{{ alicloud_secret_key }}'
-        alicloud_region: '{{ alicloud_region }}'
-        name: '{{ load_balancer_name }}'
-      register: slb_by_name
-    - debug: var=slb_by_name
+- name: Filter slb using name_regex
+  ali_slb_lb_facts:
+    name_prefix: 'ansible-slb'
+
+- name: Retrieving all slbs
+  ali_slb_lb_facts:
 '''
 
 RETURN = '''
-load_balancer_ids:
-    description: List all load balancer's id after operating slb.
+ids:
+    description: List ids of being fetched slb.
     returned: when success
     type: list
-    sample: [ "lb-dj1oi1h5l74hg22gsnugf", "lb-dj1t1xwn0y9zcr90e52i2" ]
+    sample: ["lb-dj1oi1h5l74hg22gsnugf", "lb-dj1t1xwn0y9zcr90e52i2"]
 load_balancers:
-    description: Details about the server load balancer's that were created.
-    returned: when success
-    type: list
-    sample: [
-        {
-            "address": "47.95.95.14",
-            "address_type": "internet",
-            "bandwidth": null,
-            "create_time": "2017-10-31T14:44Z",
-            "internet_charge_type": "4",
-            "load_balancer_id": "lb-dj1oi1h5l74hg22gsnugf",
-            "load_balancer_name": "demo_slb2",
-            "load_balancer_status": "active",
-            "master_zone_id": "cn-beijing-a",
-            "network_type": "classic",
-            "pay_type": "PayOnDemand",
-            "region_id": "cn-beijing",
-            "slave_zone_id": "cn-beijing-b",
-            "vpc_id": "",
-            "vswitch_id": ""
-        },
-        {
-            "address": "47.95.161.24",
-            "address_type": "internet",
-            "bandwidth": null,
-            "create_time": "2017-10-31T13:54Z",
-            "internet_charge_type": "4",
-            "load_balancer_id": "lb-dj1t1xwn0y9zcr90e52i2",
-            "load_balancer_name": "demo_slb",
-            "load_balancer_status": "active",
-            "master_zone_id": "cn-beijing-a",
-            "network_type": "classic",
-            "pay_type": "PayOnDemand",
-            "region_id": "cn-beijing",
-            "slave_zone_id": "cn-beijing-b",
-            "vpc_id": "",
-            "vswitch_id": ""
-        }
-    ]
-total:
-    description: The number of all load balancer's after operating slb.
-    returned: when success
-    type: int
-    sample: 2
+    description:
+      - info about the server load balancer that was created or deleted.
+    returned: on present
+    type: complex
+    contains:
+        address:
+            description: The IP address of the loal balancer
+            returned: always
+            type: string
+            sample: "47.94.26.126"
+        address_ipversion:
+            description: The IP address version. IPV4 or IPV6.
+            returned: always
+            type: string
+            sample: "ipv4"
+        address_type:
+            description: The load balancer internet type
+            returned: always
+            type: string
+            sample: "internet"
+        backend_servers:
+            description: The load balancer's backend servers
+            returned: always
+            type: complex
+            contains:
+                server_id:
+                    description: The backend server id
+                    returned: always
+                    type: string
+                    sample: "i-vqunci342"
+                weight:
+                    description: The backend server weight
+                    returned: always
+                    type: int
+                    sample: 100
+                description:
+                    description: The backend server description
+                    returned: always
+                    type: string
+                    sample: ""
+                type:
+                    description: The backend server type, ecs or eni
+                    returned: always
+                    type: string
+                    sample: "ecs"
+        bandwidth:
+            description: The load balancer internet bandwidth
+            returned: always
+            type: int
+            sample: 5
+        create_time:
+            description: The time of the load balancer was created
+            returned: always
+            type: string
+            sample: "2019-01-02T02:37:41Z"
+        end_time:
+            description: The time of the load balancer will be released
+            returned: always
+            type: string
+            sample: "2999-09-08T16:00:00Z"
+        id:
+            description: The ID of the load balancer was created. Same as load_balancer_id.
+            returned: always
+            type: string
+            sample: "lb-2zea9ohgtf"
+        internet_charge_type:
+            description: The load balancer internet charge type
+            returned: always
+            type: string
+            sample: "PayByTraffic"
+        listeners:
+            description: The listeners of the load balancer.
+            returned: always
+            type: complex
+            contains:
+                listener_port:
+                    description: The front-end port of the listener that is used to receive incoming traffic and
+                     distribute the traffic to the backend servers.
+                    returned: always
+                    type: int
+                    sample: 22
+                listener_protocol:
+                    description: The frontend protocol used by the SLB instance.
+                    returned: always
+                    type: string
+                    sample: tcp
+                listener_forward:
+                    description: Whether to enable listener forwarding.
+                    returned: always
+                    type: string
+                    sample: ""
+                forward_port:
+                    description: The destination listening port. It must be an existing HTTPS listening port.
+                    returned: always
+                    type: int
+                    sample: 20
+        load_balancer_id:
+            description: The ID of the load balancer was created.
+            returned: always
+            type: string
+            sample: "lb-2zea9ohgtf"
+        load_balancer_name:
+            description: The name of the load balancer was created.
+            returned: always
+            type: string
+            sample: "ansible-ali_slb_lb"
+        load_balancer_status:
+            description: The load balancer current status.
+            returned: always
+            type: string
+            sample: "active"
+        master_zone_id:
+            description: The ID of the primary zone.
+            returned: always
+            type: string
+            sample: "cn-beijing-a"
+        name:
+            description: The name of the load balancer was created.
+            returned: always
+            type: string
+            sample: "ansible-ali_slb_lb"
+        network_type:
+            description: The network type of the load balancer was created.
+            returned: always
+            type: string
+            sample: "classic"
+        pay_type:
+            description: The load balancer instance charge type.
+            returned: always
+            type: string
+            sample: "PostPaid"
+        resource_group_id:
+            description: The resource group of the load balancer belongs.
+            returned: always
+            type: string
+            sample: "rg-acfmwvvtg5owavy"
+        slave_zone_id:
+            description: The ID of the backup zone
+            returned: always
+            type: string
+            sample: "cn-beijing-d"
+        tags:
+            description: The load balancer tags
+            returned: always
+            type: complex
+            sample: {}
+        vpc_id:
+            description: The vpc of the load balancer belongs.
+            returned: always
+            type: string
+            sample: "vpc-fn3nc3"
+        vswitch_id:
+            description: The vswitch of the load balancer belongs.
+            returned: always
+            type: string
+            sample: "vsw-c3nc3r"
 '''
 
 from ansible.module_utils.basic import AnsibleModule
-from ansible.module_utils.alicloud_ecs import get_acs_connection_info, ecs_argument_spec, slb_connect
+from ansible.module_utils.alicloud_ecs import ecs_argument_spec, slb_connect
 
-HAS_ECS = False
 HAS_FOOTMARK = False
 
 try:
@@ -161,37 +252,13 @@ except ImportError:
     HAS_FOOTMARK = False
 
 
-def get_info(lb_obj):
-    """
-    get info from lb object
-    :param lb_obj: lb obj
-    :return: info of lb
-    """
-    return dict(
-        load_balancer_id=lb_obj.load_balancer_id,
-        load_balancer_name=lb_obj.load_balancer_name,
-        address=lb_obj.address,
-        internet_charge_type=lb_obj.internet_charge_type,
-        bandwidth=lb_obj.bandwidth,
-        load_balancer_status=lb_obj.load_balancer_status,
-        network_type=lb_obj.network_type,
-        master_zone_id=lb_obj.master_zone_id,
-        region_id=lb_obj.region_id,
-        create_time=lb_obj.create_time,
-        pay_type=lb_obj.pay_type,
-        slave_zone_id=lb_obj.slave_zone_id,
-        address_type=lb_obj.address_type,
-        vpc_id=lb_obj.vpc_id,
-        resource_group_id=lb_obj.resource_group_id,
-        vswitch_id=lb_obj.vswitch_id
-    )
-
-
 def main():
     argument_spec = ecs_argument_spec()
     argument_spec.update(dict(
         load_balancer_name=dict(type='list', aliases=['name']),
-        load_balancer_ids=dict(type='list', aliases=['ids'])
+        load_balancer_ids=dict(type='list', aliases=['ids']),
+        name_prefix=dict(type='str'),
+        filters=dict(type='dict')
     ))
 
     module = AnsibleModule(argument_spec=argument_spec)
@@ -199,42 +266,44 @@ def main():
     if HAS_FOOTMARK is False:
         module.fail_json(msg="Package 'footmark' required for this module.")
 
-    load_balancer_ids = module.params['load_balancer_ids']
-    load_balancer_names = module.params['load_balancer_name']
-    result = []
+    lb_ids = module.params['load_balancer_ids']
+    if not lb_ids:
+        lb_ids = []
+    name_prefix = module.params['name_prefix']
+    filters = module.params['filters']
+    if not filters:
+        filters = {}
+    for key, value in filters.items():
+        if key in ["LoadBalancerId", "load-balancer-id", "load_balancer_id"] and value not in lb_ids:
+            lb_ids.append(value)
+    lbs = []
     ids = []
 
     try:
         slb = slb_connect(module)
+        if len(lb_ids) > 0:
+            for index in range(0, len(lb_ids), 10):
+                ids_tmp = lb_ids[index:index + 10]
+                if not ids_tmp:
+                    break
+                id_str = ""
+                for id in ids_tmp:
+                    id_str += id + ','
+                filters['load_balancer_id'] = id_str[:-1]
 
-        if load_balancer_ids and (not isinstance(load_balancer_ids, list) or len(load_balancer_ids)) < 1:
-            module.fail_json(msg='load_balancer_ids should be a list of load balancer ids, aborting')
-
-        if load_balancer_names and (not isinstance(load_balancer_names, list) or len(load_balancer_names)) < 1:
-            module.fail_json(msg='load_balancer_name should be a list of load balancer name, aborting')
-
-        # list load balancers by name
-        if load_balancer_names:
-            result = []
-            ids = []
-            for  load_balancer_name in load_balancer_names:
-                for slb_res in slb.describe_load_balancers(load_balancer_name=load_balancer_name):
-                    result.append(get_info(slb_res))
-                    ids.append(slb_res.load_balancer_id)
-
-        # list load balancers by ids
-        elif load_balancer_ids:
-            slb_lb_ids = ",".join(load_balancer_ids)
-            for slb_res in slb.describe_load_balancers(load_balancer_id=slb_lb_ids):
-                result.append(get_info(slb_res))
-                ids.append(slb_res.load_balancer_id)
-        # list all load balancers
+                for lb in slb.describe_load_balancers(**filters):
+                    if name_prefix and not str(lb.load_balancer_name).startswith(name_prefix):
+                        continue
+                    lbs.append(lb.read())
+                    ids.append(lb.load_balancer_id)
         else:
-            for slb_res in slb.describe_load_balancers():
-                result.append(get_info(slb_res))
-                ids.append(slb_res.load_balancer_id)
+            for lb in slb.describe_load_balancers(**filters):
+                if name_prefix and not str(lb.load_balancer_name).startswith(name_prefix):
+                    continue
+                lbs.append(lb.read())
+                ids.append(lb.load_balancer_id)
 
-        module.exit_json(changed=False, load_balancer_ids=ids, load_balancers=result, total=len(result))
+        module.exit_json(changed=False, load_balancers=lbs, ids=ids)
     except Exception as e:
         module.fail_json(msg="Unable to describe server load balancers, and got an error: {0}.".format(e))